Changed to represent freq by int16_t from uint8_t. Now freq < 0 stands for 'not a note' freq should be represented now from 0 to NOTERANGE-1 Change how cn command sends freq Improved gui TODO check interpolating of freq when note jumps

+++ embeddedcommon/embeddednf.h
@@ -100,7 +100,7 @@ extern uint16_t folded_bins[]; //[FIXBPERO] <- The folded fourier output.
 
 //frequency of note; Note if it is == 255, then it means it is not set. It is
 //generally a value from 
-extern uint8_t  note_peak_freqs[]; //[MAXNOTES]
+extern int16_t  note_peak_freqs[]; //[MAXNOTES]
 extern uint16_t note_peak_amps[];  //[MAXNOTES] 
 extern uint16_t note_peak_amps2[]; //[MAXNOTES]  (Responds quicker)
 extern uint8_t  note_jumped_to[];  //[MAXNOTES] When a note combines into another one,
